refactor: extract project modals into separate components
- Create ProjectEditModal component for editing existing projects
- Create ProjectCreateModal component with wizard interface for new projects
- Remove embedded modal code from ProjectSelectorPanel
- All modals now consistently use proper Modal component from ui/Modal
- Improves separation of concerns and maintainability
